ER-681 move Legal Entity Mapper and Employer Account service in EmployerAccount folder.
