+++ src/Shared/Recruit.Vacancies.Client/Infrastructure/Services/EmployerAccount/IEmployerAccountService.cs
@@ -1,8 +1,8 @@
-﻿using SFA.DAS.EAS.Account.Api.Types;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Threading.Tasks;
+using SFA.DAS.EAS.Account.Api.Types;
 
-namespace Esfa.Recruit.Vacancies.Client.Infrastructure.Services
+namespace Esfa.Recruit.Vacancies.Client.Infrastructure.Services.EmployerAccount
 {
     public interface IEmployerAccountService
     {
